+++ src/main/java/chess/controller/ChessController.java
@@ -1,6 +1,6 @@
 package chess.controller;
 
-import chess.domain.Board;
+import chess.domain.board.Board;
 import chess.view.InputView;
 import chess.view.OutputView;
